added testsuite to quicktest examples renamed test_full_connection -> test_external (api vs. real srv/cli) changed test_external:do_test to accept kwargs added wrapper proc classes (helper)
